Initial script for parking report
Initial version of the Powershell script to generate a report from the EVT file. It includes the core logic for matching entry (type 1) and exit (type 2) events 169 based on the ISO card number. The script calculates parking duration and handles incomplete transactions by adding them to the report without exit data. The final output is exported to a basic Excel file.